order_mgmt/requirements.txt
Django
djangorestframework
python-dotenv
requests
requests-oauthlib
djangorestframework-simplejwt
